Abslan Laverde 05/25/2019 3:04 PM
Made adjustments to back end, gig-repository, to correctly update a Gig.  Because of the gigBands reference, merging a gig was producing an error due to new updated gig losing it's reference to the bandGigs table.  Corrected this by setting the updatedGig's gigBand reference by getting the original gig's reference and assigning it in the repository before sending the update merge.  Commiting before completing front end adjustments.